Fix plover_plugins install command
