+++ setup.cfg
@@ -58,6 +58,7 @@ packages =
 [options.entry_points]
 console_scripts =
 	plover              = plover.scripts.main:main
+	plover_plugins      = plover.plugins_manager.__main__:main
 	plover_send_command = plover.scripts.send_command:main
 plover.command =
 	set_config = plover.command.set_config:set_config
